+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 #### 3.0.0-alpha012 - 08.01.2016
 * Allow to reference git repositories - http://fsprojects.github.io/Paket/git-dependencies.html
+* Allow to use git repositories as NuGet source - http://fsprojects.github.io/Paket/git-dependencies.html#Using-Git-repositories-as-NuGet-source
 
 #### 2.41.1 - 08.01.2016
 * BUGFIX: NuGetV2: avoid revealing password also if more than one source is defined - https://github.com/fsprojects/Paket/pull/1357